Implement localisation (#328)
Since #308 was done, we now have different language files, but these
were not actually used yet, other than the English file. In this pull
request I've added actual support for localisation.

- I had to swap out the Gatsby Head API for React Helmet, since the Head
API is not wrapped in the page context and does not have access to the
locale data. This is a known issue in Gatsby, and the maintainers
recommend using Helmet.
- I've moved most of the stuff in the different Head components to a
shared SEO component.

## Todo

- [x] Right now all pages are deleted. We need to make sure the 404 page
isn't deleted.
- [x] Redirect based on the browser locale.

Closes #314.