Change category names (#372)
Add support for migrating between the legacy and new category names.
This effectively allows Snaps to use either a legacy category name or a
new one without breaking the app. The legacy category names are replaced
using a mapping a build time, but the keeps the URLs for the legacy
categories alive. The legacy URLs redirect to the new category page.

Also actually changes the underlying constants for the categories.

---------

Co-authored-by: Maarten Zuidhoorn <maarten@zuidhoorn.com>