Update provider detection to use `wallet_getSnaps` (#528)
This change only affects staging. It replaces `wallet_getAllSnaps` with
`wallet_getSnaps` when detecting the Ethereum provider
(`window.ethereum`). `wallet_getAllSnaps` is allowlisted based on the
origin, but the staging URL is not on this allowlist. We can use
`wallet_getSnaps` to reliably detect if the provider supports Snaps
though, and this works in the staging environment as well.